sqs_fetch/requirements.txt
boto3==1.24.5
psycopg2==2.9.9
sqlalchemy
pyyaml